Upgraded the following: - jquery 1.2.6 to jquery-1.3.2 - jquery-ui from 1.5.6 to 1.7.2 - jquery.checktree from .22 to .3b1
Themes:
The jquery-ui theme is no longer packaged with the LibX extension, it is instead loaded from a remote server. This makes it much simpler to switch themes for different editions. Currently to change the theme, prefs.xhtml must be edited.

Remote Loading of Templates:
This is implemented, but not used in this version ( templates are still being loaded from chrome )

General code cleanup