This commit reverts changes made by "8ea330a" (w.r.t setting transparency true for iframe)
The display style set for the body of popup.html by default is hidden or none.
In case of Chrome, it somehow or smart enough to overrides it when we are showing it as an iframe.
Whereas, in FF .. popup.html is hidden, even after load. Therefore, we have explicitly set content window
style to be visible